Fix 1 bug
If the user modified a habit for the first time in the day, an error was raised: AttributeError: 'NoneType' object has no attribute 'replace'. In line 69, habit_record was passed to local_datetime_is_today(), but habit_record does not have a date until it is added to the database, so habit_record.local_timezone_date raised the AttributeError.